Move the relative log path test to integration test.
Refactor to use testutil.
